Allow editing datasources
Identify global / personal datasources by owner and label
Allow adding color modes
Remove colormode route and just edit in detail route
