--- app-frontend/src/app/components/datasources/datasourceItem/datasourceItem.component.js
@@ -1,14 +1,0 @@
-// Component code
-import datasourceItemTpl from './datasourceItem.html';
-
-const rfDatasourceItem = {
-    templateUrl: datasourceItemTpl,
-    controller: 'DatasourceItemController',
-    bindings: {
-        datasource: '<',
-        selected: '&',
-        onSelect: '&'
-    }
-};
-
-export default rfDatasourceItem;
