--- app-frontend/src/app/components/datasources/datasourceItem/datasourceItem.controller.js
@@ -1,18 +1,0 @@
-export default class DatasourceItemController {
-    constructor($scope, $attrs) {
-        'ngInject';
-
-        this.isSelectable = $attrs.hasOwnProperty('selectable');
-        $scope.$watch(
-            () => this.selected({project: this.project}),
-            (selected) => {
-                this.selectedStatus = selected;
-            }
-        );
-    }
-
-    toggleSelected(event) {
-        this.onSelect({project: this.project, selected: !this.selectedStatus});
-        event.stopPropagation();
-    }
-}
