--- app-frontend/src/app/pages/imports/datasources/detail/colorComposites/colorComposites.html
@@ -1,37 +1,0 @@
-<table class="table">
-  <thead>
-    <tr>
-      <th>Name</th>
-      <th>Red band</th>
-      <th>Green band</th>
-      <th>Blue band</th>
-      <th><a href="#">New</a></th>
-    </tr>
-  </thead>
-  <tbody>
-    <tr ng-repeat="(preset, val) in $parent.$ctrl.colorCompositesBuffer">
-      <td>
-        <input type="text" class="form-control" placeholder="Color mode name" value="{{val.label}}">
-      </td>
-      <td>
-        <input type="text" class="form-control" placeholder="Red Band" ng-model="val.value.redBand" ng-change="$ctrl.updateBuffer(preset, 'redBand', val.value.redBand)">
-      </td>
-      <td>
-        <input type="text" class="form-control" placeholder="Green Band" ng-model="val.value.greenBand" ng-change="$ctrl.updateBuffer(preset, 'greenBand', val.value.greenBand)">
-      </td>
-      <td>
-        <input type="text" class="form-control" placeholder="Blue Band" ng-model="val.value.blueBand" ng-change="$ctrl.updateBuffer(preset, 'blueBand', val.value.blueBand)">
-      </td>
-      <td class="text-right">
-        <button class="btn btn-small" ng-click="$ctrl.deleteFromBuffer(preset)">
-          <span class="sr-only">Delete</span>
-          <i class="icon-trash"></i>
-        </button>
-      </td>
-    </tr>
-  </tbody>
-</table>
-<div class="form-group text-right">
-  <button class="btn" ng-click="$ctrl.cancel()">Cancel</button>
-  <button class="btn btn-default" ng-click="$ctrl.save()">Save Changes</button>
-</div>
