+++ app-frontend/src/app/pages/imports/datasources/detail/detail.html
@@ -5,7 +5,7 @@
       <div class="dashboard-header">
         <h1 class="h3">{{$ctrl.datasource.name}}</h1>
         <div class="flex-fill"></div>
-        <rf-toggle value="$ctrl.isPublic" on-change="$ctrl.changeVisibility()">
+        <rf-toggle value="$ctrl.isPublic" on-change="$ctrl.changeVisibility()" ng-if="$ctrl.isOwner">
           <label>Share publicly</label>
         </rf-toggle>
         <button class="btn btn-primary" ng-click="$ctrl.openImportModal()">Start an Import</button>
@@ -16,9 +16,67 @@
           <div class="cta-flex-text">
             Create preset false color composites that will be available for all imagery from the <strong>{{$ctrl.datasource.name}}</strong> datasource.
           </div>
-          <ui-view>
-              <a class="btn btn-primary" ui-sref="imports.datasources.detail.colorComposites">Edit color modes</a>
-          </ui-view>
+          <table class="table">
+            <thead>
+              <tr>
+                <th>Name</th>
+                <th>Red band</th>
+                <th>Green band</th>
+                <th>Blue band</th>
+              </tr>
+            </thead>
+            <tbody>
+              <tr ng-repeat="(preset, val) in $ctrl.colorCompositesBuffer"
+                  ng-class="{'table-row-changed': val.changed}"
+              >
+                <td>
+                  <input type="text" class="form-control" placeholder="Color mode name"
+                         ng-model="val.label"
+                         ng-change="$ctrl.onBufferChange(preset)" ng-disabled="!$ctrl.isOwner">
+                </td>
+                <td>
+                  <input type="text" class="form-control" placeholder="Red Band"
+                         ng-model="val.value.redBand"
+                         ng-change="$ctrl.updateBuffer(preset, 'redBand', val.value.redBand)"
+                         ng-disabled="!$ctrl.isOwner">
+                </td>
+                <td>
+                  <input type="text" class="form-control" placeholder="Green Band"
+                         ng-model="val.value.greenBand"
+                         ng-change="$ctrl.updateBuffer(preset, 'greenBand', val.value.greenBand)"
+                         ng-disabled="!$ctrl.isOwner">
+                </td>
+                <td>
+                  <input type="text" class="form-control" placeholder="Blue Band"
+                         ng-model="val.value.blueBand"
+                         ng-change="$ctrl.updateBuffer(preset, 'blueBand', val.value.blueBand)"
+                         ng-disabled="!$ctrl.isOwner">
+                </td>
+                <td class="text-right" ng-show="$ctrl.isOwner">
+                  <button class="btn btn-small" ng-click="$ctrl.deleteFromBuffer(preset)">
+                    <span class="sr-only">Delete</span>
+                    <i class="icon-trash"></i>
+                  </button>
+                </td>
+              </tr>
+            </tbody>
+          </table>
+          <div class="colormode-actions">
+            <div ng-show="$ctrl.isOwner">
+              <button type="button" class="btn btn-outline"
+                      ng-click="$ctrl.addCompositeRow()">
+                + Add new color mode
+              </button>
+            </div>
+            <div ng-show="$ctrl.changedBuffer">
+              <button class="btn" ng-click="$ctrl.cancel()">
+                Cancel
+              </button>
+              <button class="btn btn-default" ng-click="$ctrl.saveColorComposites()">
+                Save Changes
+              </button>
+            </div>
+          </div>
         </rf-call-to-action-item>
       </div>
     </div>
