+++ app-frontend/src/assets/styles/sass/_imports.scss
@@ -81,7 +81,8 @@
  * * * */
 @import
   "pages/dashboard",
-  "pages/editor";
+  "pages/editor",
+  "pages/datasources";
 
 /* * * *
  * Shame
