+++ app-frontend/src/assets/styles/sass/pages/_datasources.scss
@@ -1,0 +1,27 @@
+.visibility {
+    display: inline-block;
+    margin: 0 0.5em;
+    padding: .25em 0.5em;
+    border-radius: 3px;
+    width: 5em;
+    text-transform: capitalize;
+    &.private {
+        color: $shade-light;
+        border: 1px solid $shade-light;
+    }
+    &.public {
+        color: $green;
+        border: 1px solid $green;
+    }
+    &.global {
+        color: $brand-primary;
+        border: 1px solid $brand-primary;
+    }
+    text-align: center;
+}
+
+.colormode-actions {
+    display: flex;
+    flex-direction: row;
+    justify-content: space-between;
+}
